requirements.txt
ipykernel
pandas
numpy==1.26.4
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
# catboost
pymongo
from_root
evidently
dill
pyYAML
neuro_mf
boto3
mypy-boto3-s3
botocore
fastapi
uvicorn
jinja2
python-multipart
-e .